refactor(lid): break the interior-relief import cycles
